Add button that preps the blend file for submission to a render farm. (#40)
The button creates a Python script in the blend file that registers the minimal
types and properties needed for Camera Shakify shakes to animate, and marks that
script to auto-execute on file load. That way when loading the file in a
Blender without Camera Shakify installed, the camera shakes continue to work
(although there is no good way to manipulate or change them).

This is mainly intended for submitting blend files with Camera Shakify shakes to
render farms, but it might have other uses as well.